+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/client/TracingStompHandler.java
@@ -6,6 +6,7 @@ package de.gematik.test.tiger.proxy.client;
 
 import java.lang.reflect.Type;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.messaging.simp.stomp.StompFrameHandler;
 import org.springframework.messaging.simp.stomp.StompHeaders;
 
@@ -23,24 +24,44 @@ class TracingStompHandler implements StompFrameHandler {
     public void handleFrame(StompHeaders stompHeaders, Object frameContent) {
         if (frameContent instanceof TigerTracingDto) {
             final TigerTracingDto tigerTracingDto = (TigerTracingDto) frameContent;
-            TracingMessagePair messagePair = new TracingMessagePair(remoteProxyClient);
-            messagePair.setRequest(PartialTracingMessage.builder()
-                .tracingDto(tigerTracingDto)
-                .receiver(tigerTracingDto.getSender())
-                .sender(tigerTracingDto.getReceiver())
-                .messagePair(messagePair)
-                .transmissionTime(tigerTracingDto.getRequestTransmissionTime())
-                .build());
-            messagePair.setResponse(
-                PartialTracingMessage.builder()
-                    .tracingDto(tigerTracingDto)
-                    .receiver(tigerTracingDto.getReceiver())
-                    .sender(tigerTracingDto.getSender())
-                    .messagePair(messagePair)
-                    .transmissionTime(tigerTracingDto.getResponseTransmissionTime())
-                    .build());
-            remoteProxyClient.initOrUpdateMessagePart(tigerTracingDto.getRequestUuid(), messagePair.getRequest());
-            remoteProxyClient.initOrUpdateMessagePart(tigerTracingDto.getResponseUuid(), messagePair.getResponse());
+            if (StringUtils.isEmpty(tigerTracingDto.getResponseUuid())) {
+                registerNewIsolaniMessage(tigerTracingDto);
+            } else {
+                registerNewMessagePair(tigerTracingDto);
+            }
         }
+    }
+
+    private void registerNewIsolaniMessage(TigerTracingDto tigerTracingDto) {
+        var isolani = new TracingMessageIsolani(remoteProxyClient);
+        isolani.setMessage(PartialTracingMessage.builder()
+            .tracingDto(tigerTracingDto)
+            .receiver(tigerTracingDto.getSender())
+            .sender(tigerTracingDto.getReceiver())
+            .messageFrame(isolani)
+            .transmissionTime(tigerTracingDto.getRequestTransmissionTime())
+            .build());
+        remoteProxyClient.initOrUpdateMessagePart(tigerTracingDto.getRequestUuid(), isolani.getMessage());
+    }
+
+    private void registerNewMessagePair(TigerTracingDto tigerTracingDto) {
+        TracingMessagePair messagePair = new TracingMessagePair(remoteProxyClient);
+        messagePair.setRequest(PartialTracingMessage.builder()
+            .tracingDto(tigerTracingDto)
+            .receiver(tigerTracingDto.getSender())
+            .sender(tigerTracingDto.getReceiver())
+            .messageFrame(messagePair)
+            .transmissionTime(tigerTracingDto.getRequestTransmissionTime())
+            .build());
+        messagePair.setResponse(
+            PartialTracingMessage.builder()
+                .tracingDto(tigerTracingDto)
+                .receiver(tigerTracingDto.getReceiver())
+                .sender(tigerTracingDto.getSender())
+                .messageFrame(messagePair)
+                .transmissionTime(tigerTracingDto.getResponseTransmissionTime())
+                .build());
+        remoteProxyClient.initOrUpdateMessagePart(tigerTracingDto.getRequestUuid(), messagePair.getRequest());
+        remoteProxyClient.initOrUpdateMessagePart(tigerTracingDto.getResponseUuid(), messagePair.getResponse());
     }
 }