+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/TigerProxyMeshTest.java
@@ -9,16 +9,20 @@ import static org.awaitility.Awaitility.await;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.testenvmgr.junit.TigerTest;
+import java.net.ServerSocket;
+import java.net.Socket;
+import java.nio.charset.StandardCharsets;
 import java.util.concurrent.TimeUnit;
 import kong.unirest.Unirest;
 import lombok.Getter;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
 @Getter
-public class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
+class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
 
     @Test
     @TigerTest(tigerYaml = "tigerProxy:\n"
@@ -52,8 +56,8 @@ public class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
         + "      adminPort: ${free.port.4}\n"
         + "      proxiedServer: winstone\n"
         + "      proxyPort: ${free.port.5}\n")
-    @Disabled("deactivated due to buildserver problems") // TODO TGR-794
-    public void aggregateFromOneRemoteProxy(TigerTestEnvMgr envMgr) {
+//    @Disabled("deactivated due to buildserver problems") // TODO TGR-794
+    void aggregateFromOneRemoteProxy(TigerTestEnvMgr envMgr) {
         final String path = "/foobarschmar";
         Unirest.get("http://localhost:" + TigerGlobalConfiguration.readString("free.port.5") + path)
             .asString()
@@ -67,6 +71,7 @@ public class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
                     .map(p -> p.endsWith(path))
                     .orElse(false));
     }
+
     @Test
     @TigerTest(tigerYaml = "tigerProxy:\n"
         + "  skipTrafficEndpointsSubscription: true\n"
@@ -106,7 +111,7 @@ public class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
         + "      adminPort: ${free.port.6}\n"
         + "      proxiedServer: winstone\n"
         + "      proxyPort: ${free.port.7}\n")
-    @Disabled("deactivated due to buildserver problems") // TODO TGR-794
+    //@Disabled("deactivated due to buildserver problems") // TODO TGR-794
     public void testWithMultipleUpstreamProxies(TigerTestEnvMgr envMgr) {
         assertThat(envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageHistory())
             .isEmpty();
@@ -123,5 +128,82 @@ public class TigerProxyMeshTest extends AbstractTestTigerTestEnvMgr {
         await().atMost(10, TimeUnit.SECONDS)
             .until(() ->
                 envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageHistory().size() >= 4);
+    }
+
+    @SneakyThrows
+    @Test
+    @TigerTest(tigerYaml = "tigerProxy:\n"
+        + "  skipTrafficEndpointsSubscription: true\n"
+        + "  trafficEndpoints:\n"
+        + "    - http://localhost:${free.port.2}\n"
+        + "servers:\n"
+        + "  aggregatingProxy:\n"
+        + "    type: tigerProxy\n"
+        + "    dependsUpon: reverseProxy\n"
+        + "    tigerProxyCfg:\n"
+        + "      adminPort: ${free.port.2}\n"
+        + "      proxyPort: ${free.port.3}\n"
+        + "      activateRbelParsing: false\n"
+        + "      rbelBufferSizeInMb: 0\n"
+        + "      trafficEndpoints:\n"
+        + "        - http://localhost:${free.port.4}\n"
+        + "  reverseProxy:\n"
+        + "    type: tigerProxy\n"
+        + "    tigerProxyCfg:\n"
+        + "      adminPort: ${free.port.4}\n"
+        + "      proxyPort: ${free.port.5}\n"
+        + "      directReverseProxy:\n"
+        + "        hostname: localhost\n"
+        + "        port: ${free.port.6}\n")
+    void testDirectReverseProxyMeshSetup_withoutResponse(TigerTestEnvMgr envMgr) {
+        try (Socket clientSocket = new Socket("localhost",
+            TigerGlobalConfiguration.readIntegerOptional("free.port.5").get());
+            ServerSocket serverSocket = new ServerSocket(TigerGlobalConfiguration.readIntegerOptional("free.port.6").get())) {
+
+            clientSocket.getOutputStream().write("{\"foo\":\"bar\"}".getBytes());
+            clientSocket.getOutputStream().flush();
+            serverSocket.accept();
+
+            await().atMost(10, TimeUnit.SECONDS)
+                .until(() ->
+                    envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageHistory().size() >= 1);
+        }
+    }
+
+    @SneakyThrows
+    @Test
+    @TigerTest(tigerYaml = "tigerProxy:\n"
+        + "  skipTrafficEndpointsSubscription: true\n"
+        + "  trafficEndpoints:\n"
+        + "    - http://localhost:${free.port.2}\n"
+        + "servers:\n"
+        + "  aggregatingProxy:\n"
+        + "    type: tigerProxy\n"
+        + "    dependsUpon: reverseProxy\n"
+        + "    tigerProxyCfg:\n"
+        + "      adminPort: ${free.port.2}\n"
+        + "      proxyPort: ${free.port.3}\n"
+        + "      activateRbelParsing: false\n"
+        + "      rbelBufferSizeInMb: 0\n"
+        + "      trafficEndpoints:\n"
+        + "        - http://localhost:${free.port.4}\n"
+        + "  reverseProxy:\n"
+        + "    type: tigerProxy\n"
+        + "    tigerProxyCfg:\n"
+        + "      adminPort: ${free.port.4}\n"
+        + "      proxyPort: ${free.port.5}\n"
+        + "      directReverseProxy:\n"
+        + "        hostname: localhost\n"
+        + "        port: ${free.port.2}\n")
+    void testDirectReverseProxyMeshSetup_withResponse(TigerTestEnvMgr envMgr) {
+        try (Socket clientSocket = new Socket("127.0.0.1",
+            Integer.parseInt(TigerGlobalConfiguration.resolvePlaceholders("${free.port.5}")))) {
+
+            clientSocket.getOutputStream().write("{\"foo\":\"bar\"}".getBytes());
+            clientSocket.getOutputStream().flush();
+            await().atMost(10, TimeUnit.SECONDS)
+                .until(() ->
+                    envMgr.getLocalTigerProxyOrFail().getRbelLogger().getMessageHistory().size() >= 1);
+        }
     }
 }